+++ test/unit/test_iqm_api_config.cpp
@@ -32,6 +32,8 @@ TEST(APIConfigTest, UnifiedApiUrl) {
       "http://test.url/api/v1/jobs/123/artifacts/measurement_counts");
   EXPECT_EQ(config.url(iqm::API_ENDPOINT::GET_JOB_ARTIFACT_MEASUREMENTS, "123"),
             "http://test.url/api/v1/jobs/123/artifacts/measurements");
+  EXPECT_EQ(config.url(iqm::API_ENDPOINT::GET_JOB_PAYLOAD, "123"),
+            "http://test.url/api/v1/jobs/123/payload");
   EXPECT_EQ(config.url(iqm::API_ENDPOINT::CANCEL_JOB, "123"),
             "http://test.url/api/v1/jobs/123/cancel");
   EXPECT_EQ(config.url(iqm::API_ENDPOINT::GET_QUANTUM_COMPUTERS),